GeoPress/requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
Pillow==9.0.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.0
